explorer: consolidate the snapshot-param module (code review finding A2)
apps/explorer/lib/snapshot-param.ts used to be a thin re-export of
platform/lib/snapshot-param.ts (plus isSha/shortSha/snapshotSrc/
snapshotListing, none of which have any apps/ dependency), each with its own
test file — two names for the same rule invited exactly the "which one do I
edit" question a later change would eventually get wrong.

Moved every function into platform/lib/snapshot-param.ts, merged the two test
files into platform/lib/snapshot-param.test.ts, and deleted the apps/explorer
copy. Preview.tsx and Listing.tsx now import directly from
@platform/lib/snapshot-param — the same direction router.ts already takes.
