+++ frontend/src/platform/lib/snapshot-param.test.ts
@@ -3,13 +3,18 @@ import {
   carries,
   getResolvedSnapshot,
   getSnapshotAppDir,
+  isSha,
   rewriteSnapshotPath,
   setResolvedSnapshot,
+  shortSha,
+  snapshotListing,
+  snapshotSrc,
 } from "@platform/lib/snapshot-param";
 
 const APP = "/repo/myapp";
 const DIR = "/cache/key/abc1234";
 const SNAP = { sha: "abc1234", dir: DIR, app_dir: APP };
+const SHA = "1f0c3a9e2b7d4c5f6a8b9c0d1e2f3a4b5c6d7e8f";
 
 describe("carries — the app-folder-scoped carry table", () => {
   it("carries a hop to the same app folder", () => {
@@ -81,5 +86,69 @@ describe("rewriteSnapshotPath — mirrors static/runtime.js's rewritePath", () =
   it("rewrites nothing with no active snapshot", () => {
     setResolvedSnapshot(null);
     expect(rewriteSnapshotPath(APP + "/reader.py")).toBe(APP + "/reader.py");
+  });
+});
+
+describe("isSha", () => {
+  it("takes a hex object name, full or abbreviated", () => {
+    expect(isSha(SHA)).toBe(true);
+    expect(isSha("1f0c3a9")).toBe(true);
+  });
+
+  // Ported from preview-rev.test.ts's junk-value case: the same shape must
+  // refuse the same junk, since both modules validate a value bound for the
+  // same server endpoint's `sha` argv position.
+  it("reads anything that is not a sha as junk", () => {
+    for (const bad of [null, undefined, "", "HEAD", "HEAD~2", "../etc/passwd",
+                       "1f0", "zzzz", 12345, {}, "1f0c3a9 --upload-pack=x"]) {
+      expect(isSha(bad)).toBe(false);
+    }
+  });
+});
+
+describe("shortSha", () => {
+  it("takes the same seven-character abbreviation the git template shows", () => {
+    expect(shortSha(SHA)).toBe("1f0c3a9");
+  });
+});
+
+describe("snapshotSrc", () => {
+  it("appends _snapshot to a real src", () => {
+    expect(snapshotSrc("/render?path=x", "abc1234")).toBe("/render?path=x&_snapshot=abc1234");
+  });
+
+  it("passes a null src or a null sha through unchanged", () => {
+    expect(snapshotSrc(null, "abc1234")).toBe(null);
+    expect(snapshotSrc("/render?path=x", null)).toBe("/render?path=x");
+  });
+});
+
+describe("snapshotListing", () => {
+  it("is not in-snapshot with no active resolution", () => {
+    setResolvedSnapshot(null);
+    expect(snapshotListing(APP)).toEqual({ inSnapshot: false, listPath: APP });
+  });
+
+  it("rewrites the listing target when the folder is inside the app", () => {
+    setResolvedSnapshot(SNAP);
+    expect(snapshotListing(APP)).toEqual({ inSnapshot: true, listPath: DIR });
+    expect(snapshotListing(APP + "/sub")).toEqual({
+      inSnapshot: true,
+      listPath: DIR + "/sub",
+    });
+    setResolvedSnapshot(null);
+  });
+
+  it("passes through unrewritten when the folder is outside the app", () => {
+    setResolvedSnapshot(SNAP);
+    expect(snapshotListing("/repo/otherapp")).toEqual({
+      inSnapshot: false,
+      listPath: "/repo/otherapp",
+    });
+    expect(snapshotListing("/repo/myapp-notes")).toEqual({
+      inSnapshot: false,
+      listPath: "/repo/myapp-notes",
+    });
+    setResolvedSnapshot(null);
   });
 });
